rename tests from 'IntegrationTest' to 'JarTest'
