+++ tests/org.eclipse.n4js.ide.integrationtests/META-INF/MANIFEST.MF
@@ -3,8 +3,8 @@ Bundle-ManifestVersion: 2
 Bundle-Name: %pluginName
 Bundle-Vendor: %providerName
 Bundle-Version: 0.0.1.qualifier
-Bundle-SymbolicName: org.eclipse.n4js.ide.integrationtests
-Export-Package: org.eclipse.n4js.ide.integrationtests
+Bundle-SymbolicName: org.eclipse.n4js.ide.jartests
+Export-Package: org.eclipse.n4js.ide.jartests
 Require-Bundle: org.eclipse.n4js,
  org.eclipse.n4js.utils,
  org.junit,
@@ -18,4 +18,4 @@ Require-Bundle: org.eclipse.n4js,
  org.eclipse.n4js.cli.tests.helper,
  org.eclipse.n4js.smith
 Bundle-RequiredExecutionEnvironment: JavaSE-11
-Automatic-Module-Name: org.eclipse.n4js.ide.integrationtests
+Automatic-Module-Name: org.eclipse.n4js.ide.jartests
